Pokedex Simulator -Samples random pokemon

Pokedex Simulator -Samples random pokemon

Usage

pokedex.sim(n, replace = T, gen = 1, prob)

Arguments

n

number of pokemon to return

replace

controls whether to sample with replacement or not

gen

controls which generation of pokemon to sample from, default is 1st gen

prob

an optional parameter that controls probability of sampling see sample()

Value

a dataframe of random selected pokemon

Examples

pokedex.sim(10)
#>       X pokedex_number             name generation status type_number  type_1
#> 81   80             62        Poliwrath          1 Normal           2   Water
#> 101 100             77  Galarian Ponyta          1 Normal           1 Psychic
#> 122 121             93          Haunter          1 Normal           2   Ghost
#> 133 132            103        Exeggutor          1 Normal           2   Grass
#> 153 152            119          Seaking          1 Normal           1   Water
#> 40   39             28 Alolan Sandslash          1 Normal           2     Ice
#> 172 171            134         Vaporeon          1 Normal           1   Water
#> 121 120             92           Gastly          1 Normal           2   Ghost
#> 134 133            103 Alolan Exeggutor          1 Normal           2   Grass
#> 166 165            130         Gyarados          1 Normal           2   Water
#>       type_2 height_m weight_kg abilities_number    ability_1   ability_2
#> 81  Fighting      1.3      54.0                3 Water Absorb        Damp
#> 101               0.8      24.0                3     Run Away Pastel Veil
#> 122   Poison      1.6       0.1                1     Levitate            
#> 133  Psychic      2.0     120.0                2  Chlorophyll            
#> 153               1.3      39.0                3   Swift Swim  Water Veil
#> 40     Steel      1.2      55.0                2   Snow Cloak            
#> 172               1.0      29.0                2 Water Absorb            
#> 121   Poison      1.3       0.1                1     Levitate            
#> 134   Dragon     10.9     415.6                2        Frisk            
#> 166   Flying      6.5     235.0                2   Intimidate            
#>     ability_hidden total_points  hp attack defense sp_attack sp_defense speed
#> 81      Swift Swim          510  90     95      95        70         90    70
#> 101   Anticipation          410  50     85      55        65         65    90
#> 122                         405  45     50      45       115         55    95
#> 133        Harvest          530  95     95      85       125         75    55
#> 153  Lightning Rod          450  80     92      65        65         80    68
#> 40      Slush Rush          450  75    100     120        25         65    65
#> 172      Hydration          525 130     65      60       110         95    65
#> 121                         310  30     35      30       100         35    80
#> 134        Harvest          530  95    105      85       125         75    45
#> 166          Moxie          540  95    125      79        60        100    81
#>     catch_rate base_friendship base_experience growth_rate egg_type_number
#> 81          45              70             230 Medium Slow               1
#> 101         NA              NA              NA Medium Fast               1
#> 122         90              70             142 Medium Slow               1
#> 133         45              70             186        Slow               1
#> 153         60              70             158 Medium Fast               1
#> 40          90              70             158 Medium Fast               1
#> 172         45              70             184 Medium Fast               1
#> 121        190              70              62 Medium Slow               1
#> 134         45              70             186        Slow               1
#> 166         45              70             189        Slow               2
#>     egg_type_1 egg_type_2 percentage_male egg_cycles against_normal
#> 81     Water 1                       50.0         20            1.0
#> 101      Field                         NA         20            1.0
#> 122  Amorphous                       50.0         20            0.0
#> 133      Grass                       50.0         20            1.0
#> 153    Water 2                       50.0         20            1.0
#> 40       Field                       50.0         20            0.5
#> 172      Field                       87.5         35            1.0
#> 121  Amorphous                       50.0         20            0.0
#> 134      Grass                       50.0         20            1.0
#> 166     Dragon    Water 2            50.0          5            1.0
#>     against_fire against_water against_electric against_grass against_ice
#> 81           0.5          0.00             2.00          2.00        0.50
#> 101          0.0          2.00             1.00          0.50        0.50
#> 122          1.0          1.00             1.00          0.50        1.00
#> 133          2.0          0.50             0.50          0.50        2.00
#> 153          0.5          0.50             2.00          2.00        0.50
#> 40           4.0          1.00             1.00          0.50        0.25
#> 172          0.5          0.00             2.00          2.00        0.50
#> 121          1.0          1.00             1.00          0.50        1.00
#> 134          1.0          0.25             0.25          0.25        4.00
#> 166          0.5          0.50             4.00          1.00        1.00
#>     against_fight against_poison against_ground against_flying against_psychic
#> 81            1.0           1.00            1.0            2.0             2.0
#> 101           1.0           1.00            2.0            1.0             1.0
#> 122           0.0           0.25            0.0            1.0             2.0
#> 133           0.5           2.00            0.5            2.0             0.5
#> 153           1.0           1.00            1.0            1.0             1.0
#> 40            4.0           0.00            2.0            0.5             0.5
#> 172           1.0           1.00            1.0            1.0             1.0
#> 121           0.0           0.25            0.0            1.0             2.0
#> 134           1.0           2.00            0.5            2.0             1.0
#> 166           0.5           1.00            0.0            1.0             1.0
#>     against_bug against_rock against_ghost against_dragon against_dark
#> 81         0.50          0.5             1            1.0          0.5
#> 101        0.50          2.0             1            1.0          1.0
#> 122        0.25          1.0             2            1.0          2.0
#> 133        4.00          1.0             2            1.0          2.0
#> 153        1.00          1.0             1            1.0          1.0
#> 40         0.50          1.0             1            0.5          1.0
#> 172        1.00          1.0             1            1.0          1.0
#> 121        0.25          1.0             2            1.0          2.0
#> 134        2.00          1.0             1            2.0          1.0
#> 166        0.50          2.0             1            1.0          1.0
#>     against_steel against_fairy
#> 81            0.5           2.0
#> 101           0.5           0.5
#> 122           1.0           0.5
#> 133           1.0           1.0
#> 153           0.5           1.0
#> 40            1.0           0.5
#> 172           0.5           1.0
#> 121           1.0           0.5
#> 134           1.0           2.0
#> 166           0.5           1.0